requirements.txt
absl-py==0.1.10
alabaster==0.7.10
appnope==0.1.0
astroid==1.6.1
Babel==2.5.3
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
cloudpickle==0.5.2
cycler==0.10.0
decorator==4.2.1
docutils==0.14
entrypoints==0.2.3
html5lib==0.9999999
idna==2.6
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras==2.1.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
numpy==1.14.1
numpydoc==0.7.0
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.4.3
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
pyparsing==2.2.0
PyQt5==5.10
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
rope==0.10.7
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
sip==4.19.7
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.0
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
testpath==0.3.1
textblob==0.15.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wordcloud==1.3.1
wrapt==1.10.11
